# mj_web_openapi.yaml
openapi: 3.0.3
info:
  title: MJ_WEB API
  version: 1.0.0
servers:
- url: /api
paths:
...
  /account/profile:
    get:
      summary: Connector-backed profile
      responses:
        '200':
          description: OK
      security:
      - cookieAuth: []
    put:
      summary: Update local contact & auth email
      responses:
        '200':
          description: OK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      security:
      - cookieAuth: []
  /account/addresses:
    get:
      summary: Connector-backed addresses
      responses:
        '200':
          description: OK
      security:
      - cookieAuth: []
...
  /cart:
    get:
      summary: Get cart
      responses:
        '200':
          description: OK
      security:
      - cookieAuth: []
  /cart/add:
    post:
      summary: Add item
      responses:
        '200':
          description: OK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: integer
                variationId:
                  type: integer
                qty:
                  type: integer
              required:
              - productId
              - variationId
              - qty
      security:
      - cookieAuth: []
  /cart/update:
    patch:
      summary: Update line qty
      responses:
        '200':
          description: OK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                qty:
                  type: integer
              required:
              - id
              - qty
      security:
      - cookieAuth: []
  /cart/remove/{id}:
    delete:
      summary: Remove line
      responses:
        '200':
          description: OK
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      security:
      - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT